projects/ynai5-mcps/mcps/telegram-mcp/requirements.txt
fastmcp>=2.0.0
telethon>=1.36.0
python-dotenv>=1.0.0
cryptg>=0.4.0
